requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.46
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
